backend/requirements.txt
fastapi
uvicorn
pymongo
python-multipart
google-auth
requests
pymysql
cryptography